refactor(tests): Consolidate Playwright tests
This commit refactors the Playwright test suite by consolidating all tests into a single `project-health.spec.ts` file. Redundant tests have been removed, and the remaining tests have been organized into logical groups for improved readability and maintainability.

The `package.json` has been updated to remove obsolete test scripts, and the visual regression snapshots have been updated to align with the new test structure.

This change simplifies the test suite and better aligns it with the project's goal of having a robust and maintainable testing strategy.
